refactor(pipeline-email): home email pipeline stages in the package
Move summarize/extract-contacts/stats stages + EmailPipelineCtx from
stories-brain into @dxos/pipeline-email as exported stages (backend-agnostic:
summarize resolves LanguageModel from AiService). Promote the src-used deps
(@dxos/ai, @dxos/echo, @dxos/extractor-lib, @dxos/util, @effect/ai) to
dependencies. stories-brain drops the now-unneeded deps and takes
@dxos/pipeline-email to assemble the stages.
