+++ packages/core/compute/pipeline-email/package.json
@@ -26,21 +26,21 @@
     "src"
   ],
   "dependencies": {
+    "@dxos/ai": "workspace:*",
+    "@dxos/echo": "workspace:*",
+    "@dxos/extractor-lib": "workspace:*",
     "@dxos/node-std": "workspace:*",
     "@dxos/pipeline": "workspace:*",
     "@dxos/pipeline-rdf": "workspace:*",
     "@dxos/types": "workspace:*",
+    "@dxos/util": "workspace:*",
+    "@effect/ai": "catalog:",
     "hyparquet": "catalog:"
   },
   "devDependencies": {
-    "@dxos/ai": "workspace:*",
-    "@dxos/echo": "workspace:*",
     "@dxos/echo-client": "workspace:*",
     "@dxos/effect": "workspace:*",
-    "@dxos/extractor-lib": "workspace:*",
     "@dxos/log": "workspace:*",
-    "@dxos/util": "workspace:*",
-    "@effect/ai": "catalog:",
     "effect": "catalog:",
     "hyparquet-writer": "catalog:",
     "vitest": "catalog:"
