+++ packages/core/compute/pipeline-email/src/stages.ts
@@ -3,18 +3,43 @@
 //
 
 import * as LanguageModel from '@effect/ai/LanguageModel';
+import * as Context from 'effect/Context';
 import * as Effect from 'effect/Effect';
 import * as Layer from 'effect/Layer';
 
 import { AiService } from '@dxos/ai';
+import { type Database } from '@dxos/echo';
 import { extractContact } from '@dxos/extractor-lib';
 import { Stage } from '@dxos/pipeline';
 import { ContentBlock, Message } from '@dxos/types';
 import { trim } from '@dxos/util';
 
-import { EmailPipelineCtx, type Summary } from './run';
-
 const SUMMARIZE_MODEL = 'com.anthropic.model.claude-haiku-4-5.default';
+
+/** Summary produced by the summarize stage. */
+export type Summary = { readonly summary: string; readonly isSpam: boolean; readonly keywords: readonly string[] };
+
+/** Running tallies produced by the stats stage (mutable accumulator). */
+export type Stats = { from: Map<string, number>; to: Map<string, number>; total: number; spam: number };
+
+/** Per-message summary result keyed by messageId (for a summary view). */
+export type SummaryResult = { readonly summaries: ReadonlyArray<{ messageId: string; summary: Summary }> };
+
+export const emptyStats = (): Stats => ({ from: new Map(), to: new Map(), total: 0, spam: 0 });
+
+/**
+ * Shared context threaded through the email stages via Effect's Requirements channel. `db` is the
+ * ECHO space database (browser-safe); `stats` and `summaries` are mutable accumulators read after
+ * the run. Callers provide it once at the pipeline edge (`Effect.provide`).
+ */
+export class EmailPipelineCtx extends Context.Tag('@dxos/pipeline-email/EmailPipelineCtx')<
+  EmailPipelineCtx,
+  {
+    readonly db: Database.Database;
+    readonly stats: Stats;
+    readonly summaries: Array<{ messageId: string; summary: Summary }>;
+  }
+>() {}
 
 const SUMMARIZE_PROMPT = trim`
   Summarize the following email in one sentence, decide whether it is spam, and list up to five keywords.
