+++ packages/stories/stories-brain/package.json
@@ -21,10 +21,10 @@
     "@dxos/crawler": "workspace:*",
     "@dxos/echo": "workspace:*",
     "@dxos/effect": "workspace:*",
-    "@dxos/extractor-lib": "workspace:*",
     "@dxos/keys": "workspace:*",
     "@dxos/nlp": "workspace:*",
     "@dxos/pipeline": "workspace:*",
+    "@dxos/pipeline-email": "workspace:*",
     "@dxos/pipeline-rdf": "workspace:*",
     "@dxos/plugin-client": "workspace:*",
     "@dxos/plugin-discord": "workspace:*",
@@ -46,7 +46,6 @@
     "@dxos/ui-theme": "workspace:*",
     "@dxos/ui-types": "workspace:*",
     "@dxos/util": "workspace:*",
-    "@effect/ai": "catalog:",
     "effect": "catalog:"
   },
   "devDependencies": {
